wip: styling button
