+++ src/css/ifm-variables.scss
@@ -48,9 +48,7 @@
   --ifm-menu-color-background-hover: var(--bt-menu-color-background-hover);
 
   /* Search  */
-  --ifm-navbar-search-input-background-color: var(
-    --bt-sidebar-background-color
-  );
+  --ifm-navbar-search-input-background-color: var(--bt-sidebar-background-color);
   --ifm-navbar-search-input-color: var(--ifm-font-color-base);
   --ifm-navbar-search-input-placeholder-color: var(--ifm-font-color-base);
   --ifm-navbar-search-input-icon: url("data:image/svg+xml,%3Csvg width='16' height='17' viewBox='0 0 16 17' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath fill-rule='evenodd' clip-rule='evenodd' d='M11.1672 3.84858C9.01033 1.68379 5.51679 1.68379 3.35991 3.84947L3.35786 3.85152C1.19366 6.0078 1.19211 9.50217 3.35983 11.6699C5.51625 13.8263 9.01201 13.828 11.18 11.6691C13.3363 9.51191 13.336 6.01739 11.1792 3.86055C11.1773 3.85872 11.1755 3.85688 11.1737 3.85503C11.1715 3.8529 11.1694 3.85075 11.1672 3.84858ZM12.2343 2.79443C12.2328 2.79291 12.2313 2.79138 12.2298 2.78987C9.48701 0.036999 5.04132 0.0366578 2.29812 2.78995C-0.454663 5.53376 -0.452765 9.97862 2.29917 12.7306C5.04266 15.474 9.4867 15.4725 12.2387 12.7316L12.2398 12.7306C14.9827 9.98766 14.9827 5.54279 12.2398 2.79989C12.238 2.79806 12.2362 2.79624 12.2343 2.79443Z' fill='%238086A0'/%3E%3Cpath fill-rule='evenodd' clip-rule='evenodd' d='M11.1792 11.6799C11.4721 11.387 11.9469 11.387 12.2398 11.6799L15.5298 14.9699C15.8227 15.2628 15.8227 15.7377 15.5298 16.0306C15.2369 16.3234 14.7621 16.3234 14.4692 16.0306L11.1792 12.7406C10.8863 12.4477 10.8863 11.9728 11.1792 11.6799Z' fill='%238C90A4'/%3E%3C/svg%3E");
@@ -84,5 +82,12 @@
   --ifm-navbar-search-input-color: var(--ifm-font-color-base);
   --ifm-navbar-search-input-placeholder-color: var(--ifm-font-color-base);
   --ifm-navbar-search-input-icon: url("data:image/svg+xml,%3Csvg width='16' height='17' viewBox='0 0 16 17' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath fill-rule='evenodd' clip-rule='evenodd' d='M11.1672 3.84858C9.01033 1.68379 5.51679 1.68379 3.35991 3.84947L3.35786 3.85152C1.19366 6.0078 1.19211 9.50217 3.35983 11.6699C5.51625 13.8263 9.01201 13.828 11.18 11.6691C13.3363 9.51191 13.336 6.01739 11.1792 3.86055C11.1773 3.85872 11.1755 3.85688 11.1737 3.85503C11.1715 3.8529 11.1694 3.85075 11.1672 3.84858ZM12.2343 2.79443C12.2328 2.79291 12.2313 2.79138 12.2298 2.78987C9.48701 0.036999 5.04132 0.0366578 2.29812 2.78995C-0.454663 5.53376 -0.452765 9.97862 2.29917 12.7306C5.04266 15.474 9.4867 15.4725 12.2387 12.7316L12.2398 12.7306C14.9827 9.98766 14.9827 5.54279 12.2398 2.79989C12.238 2.79806 12.2362 2.79624 12.2343 2.79443Z' fill='%238086A0'/%3E%3Cpath fill-rule='evenodd' clip-rule='evenodd' d='M11.1792 11.6799C11.4721 11.387 11.9469 11.387 12.2398 11.6799L15.5298 14.9699C15.8227 15.2628 15.8227 15.7377 15.5298 16.0306C15.2369 16.3234 14.7621 16.3234 14.4692 16.0306L11.1792 12.7406C10.8863 12.4477 10.8863 11.9728 11.1792 11.6799Z' fill='%238086A0'/%3E%3C/svg%3E");
-}
 
+  /* Markdown button */
+  --ifm-color-white: #ffffff;
+  --imf-toggle-button-background-color: rgba(31, 35, 45, 0.95);
+  --imf-toggle-button-option-color-hover: rgba(209, 215, 255, 0.1);
+  --imf-toggle-button-border-color: rgba(4, 6, 17, 0.2);
+  --imf-toggle-button-border-color-active: rgba(0, 210, 239, 0.5);
+
+}
