+++ src/theme/DocItem/Header/styles.module.css
@@ -1,0 +1,114 @@
+.docHeader {
+    margin-bottom: 2rem;
+}
+
+.docHeaderContent {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    gap: 1rem;
+}
+
+.docHeaderContent h1 {
+    margin-bottom: 0;
+    flex: 1;
+}
+
+.markdown {
+    position: relative;
+    flex-shrink: 0;
+    padding-top: 0.75rem;
+}
+
+.markdownButton {
+    display: flex !important;
+    align-items: center;
+    justify-content: space-between;
+    color: var(--ifm-color-white);
+    background: var(--imf-toggle-button-background-color);
+    border: 1px solid var(--imf-toggle-button-border-color);
+    border-radius: 8px;
+    cursor: pointer;
+    padding: 8px 8px 8px 12px;
+    font: inherit;
+    font-size: 14px;
+    transition: all 0.2s ease;
+    white-space: nowrap;
+    font-weight: 500;
+}
+
+.markdownButton[aria-expanded="true"] {
+    color: var(--ifm-color-primary);
+    border-color: var(--imf-toggle-button-border-color-active);
+    background-color: transparent;
+}
+
+.markdownButton[aria-expanded="true"]:hover {
+    background-color: transparent;
+}
+
+.markdownButton[aria-expanded="true"]::after {
+    background-image: url('data:image/svg+xml,<svg width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M4 6L8 10L12 6" stroke="%2300D2EF" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"/></svg>');
+}
+
+.markdownButton::after {
+    content: '';
+    width: 16px;
+    height: 16px;
+    padding-left: 12px;
+    background-image: url('data:image/svg+xml,<svg width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M4 6L8 10L12 6" stroke="white" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"/></svg>');
+    background-repeat: no-repeat;
+    background-position: center;
+    transition: transform 0.2s ease;
+}
+
+.markdownButton:hover {
+    background: rgba(31, 35, 45, 1);
+}
+
+.dropdownContainer {
+    position: absolute;
+    top: 100%;
+    right: 0;
+    background: rgba(31, 35, 45, 0.95);
+    border: none;
+    border-radius: 8px;
+    box-shadow: 0px 8px 10px 0px rgba(4, 6, 17, 0.2),
+        0px 0px 0px 1px rgba(255, 255, 255, 0.1) inset;
+    padding: 0;
+    width: 125px;
+    z-index: 100;
+    margin-top: 4px;
+    overflow: hidden;
+}
+
+.dropdownItem {
+    display: flex;
+    align-items: center;
+    gap: 12px;
+    padding: 6px 16px 6px 12px;
+    color: var(--ifm-color-white);
+    font-size: 14px;
+    line-height: 21px;
+    text-decoration: none;
+    transition: all 0.2s ease;
+    width: 100%;
+    border: none;
+    background: none;
+    text-align: left;
+    cursor: pointer;
+    margin: 0;
+    font: inherit;
+}
+
+.dropdownItem:hover {
+    background: rgba(255, 255, 255, 0.1);
+    text-decoration: none;
+}
+
+.dropdownIcon {
+    width: 16px;
+    height: 16px;
+    opacity: 0.7;
+    fill: currentColor;
+}
