--- src/theme/TOC/styles.module.css
@@ -1,95 +1,0 @@
-.divider {
-    margin: 8px 0;
-    border-color: var(--bt-default-border-color);
-}
-
-.markdown {
-    position: relative;
-    padding: 4px var(--ifm-toc-padding-horizontal);
-    margin-top: 8px;
-}
-
-.markdownButton {
-    display: flex !important;
-    align-items: center;
-    gap: 8px;
-    color: var(--ifm-toc-link-color) !important;
-    background: none;
-    border: none;
-    cursor: pointer;
-    padding: 0;
-    width: 100%;
-    text-align: left;
-    font: inherit;
-}
-
-.markdownButton::before {
-    content: '';
-    mask: url('data:image/svg+xml,<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 16" width="16" height="16"><path d="M14.85 3H1.15C.52 3 0 3.52 0 4.15v7.69C0 12.48.52 13 1.15 13h13.69c.64 0 1.15-.52 1.15-1.15v-7.7C16 3.52 15.48 3 14.85 3zM9 11H7V8L5.5 9.92 4 8v3H2V5h2l1.5 2L7 5h2v6zm2.99.5L9.5 8H11V5h2v3h1.5l-2.51 3.5z"/></svg>') no-repeat 50% 50%;
-    -webkit-mask: url('data:image/svg+xml,<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 16" width="16" height="16"><path d="M14.85 3H1.15C.52 3 0 3.52 0 4.15v7.69C0 12.48.52 13 1.15 13h13.69c.64 0 1.15-.52 1.15-1.15v-7.7C16 3.52 15.48 3 14.85 3zM9 11H7V8L5.5 9.92 4 8v3H2V5h2l1.5 2L7 5h2v6zm2.99.5L9.5 8H11V5h2v3h1.5l-2.51 3.5z"/></svg>') no-repeat 50% 50%;
-    mask-size: cover;
-    -webkit-mask-size: cover;
-    background-color: currentColor;
-    width: 16px;
-    height: 16px;
-    display: inline-block;
-}
-
-.markdownButton:hover {
-    color: var(--ifm-heading-color);
-}
-
-.dropdownContainer {
-    position: absolute;
-    top: 100%;
-    left: 0;
-    background: var(--ifm-background-surface-color);
-    border: 1px solid var(--ifm-color-emphasis-300);
-    border-radius: 8px;
-    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
-    padding: 8px 0;
-    min-width: 180px;
-    z-index: 100;
-    margin-top: 4px;
-}
-
-.dropdownItem {
-    display: flex;
-    align-items: center;
-    gap: 8px;
-    padding: 8px 16px;
-    color: var(--ifm-color-content);
-    font-size: 14px;
-    text-decoration: none;
-    transition: background-color 0.2s;
-    width: 100%;
-    border: none;
-    background: none;
-    text-align: left;
-    cursor: pointer;
-    margin: 0;
-    font: inherit;
-}
-
-.dropdownItem:hover {
-    background-color: var(--ifm-color-emphasis-100);
-    text-decoration: none;
-    color: var(--ifm-color-primary);
-}
-
-.dropdownIcon {
-    width: 16px;
-    height: 16px;
-    opacity: 0.7;
-}
-
-/* Add a divider before the Markdown link */
-.markdown::after {
-    content: '';
-    position: absolute;
-    left: var(--ifm-toc-padding-horizontal);
-    right: var(--ifm-toc-padding-horizontal);
-    top: -8px;
-    height: 1px;
-    background-color: var(--ifm-color-emphasis-300);
-}
